Seperated GraphIO from GraphConversions. Added Documentation
