+++ documentation/mkdocs.yml
@@ -59,6 +59,7 @@ nav:
       - Examples:
          'api/examples/graph_construction_examples.md'
   - Graph Conversions: 'api/graph_conversions.md'
+  - Graph IO: 'api/graph_io.md'
   - Population Averaging: 'api/population_averaging.md'
   - Dynamic Utilities: 'api/dynamic_utils.md'
   - Graph Measures: 'api/graph_measures.md'
